// src/app/_dto/ico.dto.ts
export class IcoDto {

    coCode: string = "";

    coName: string = "";

    coShName: string = "";

    add1: string | null = "";

    add2: string | null = "";

    add3: string | null = "";

    add4: string | null = "";

    tel: string | null = "";

    fax: string | null = "";

    email: string | null = "";

    website: string | null = "";

    poLevel: string | null = "";

    poDefDv: string | null = "";

    poDefBr: string | null = "";

    dvWiseItemsYn: string = "N";

    itemDefDv: string | null = "";

    assetLevel: string | null = "";

    assetDefDv: string | null = "";

    assetDefBr: string | null = "";

    jobLevel: string | null = "";

    jobDefDv: string | null = "";

    jobDefBr: string | null = "";

    jobHeadLevel: string | null = "";

    jobHeadDefDv: string | null = "";

    jobHeadDefBr: string | null = "";

    yrSDt: Date | null = new Date();

    yrEDt: Date | null = new Date();

    plAc: string | null = "";

    pdcRecAc: string | null = "";

    pdcPayAc: string | null = "";

    baseCurCode: string | null = "";

    rateDecPts: number | null = 0;

    amtDecPts: number | null = 0;

    invGrp1Name: string | null = "";

    invGrp2Name: string | null = "";

    invGrp3Name: string | null = "";

    invGrp4Name: string | null = "";

    currencyMultiple: string = "N";

    assetsDeprInt: string | null = "";

    pdcJv: string | null = "";

    multiUnit: string = "N";

    modiCloseDate: Date = new Date();

    logData: string = "N";

    invGrp5Name: string | null = "";

    invGrp6Name: string | null = "";

    invGrp7Name: string | null = "";

    invGrp8Name: string | null = "";

    coNameAr: string | null = "";

    add1Ar: string | null = "";

    add2Ar: string | null = "";

    add3Ar: string | null = "";

    add4Ar: string | null = "";

    isArabic: string = "";

}